Combine SSL tests into one class
